+++ frontend/src/metabase-lib/queries/drills/zoom-drill.ts
@@ -1,3 +1,4 @@
+import { t } from "ttag";
 import {
   ClickObject,
   DimensionValue,
@@ -5,6 +6,8 @@ import {
 import { LocalFieldReference } from "metabase-types/types/Query";
 import { drillDownForDimensions } from "metabase-lib/queries/utils/drilldown";
 import Question from "metabase-lib/Question";
+import { FieldDimension } from "metabase-lib/Dimension";
+import { formatBucketing } from "metabase-lib/queries/utils/query-time";
 
 export const getNextZoomDrilldown = (
   question: Question,
@@ -13,6 +16,10 @@ export const getNextZoomDrilldown = (
   dimensions: DimensionValue[];
   drilldown: { breakouts: LocalFieldReference[] };
 } | null => {
+  if (!question.query().isEditable()) {
+    return null;
+  }
+
   const dimensions = clicked?.dimensions ?? [];
   const drilldown = drillDownForDimensions(dimensions, question.metadata());
 
@@ -24,6 +31,39 @@ export const getNextZoomDrilldown = (
   }
 
   return null;
+};
+
+export const getZoomDrillTitle = (
+  dimensions: DimensionValue[],
+  drilldown: { breakouts: LocalFieldReference[] },
+): string => {
+  let currentGranularity;
+  dimensions.some(dimension => {
+    if (dimension.column.field_ref) {
+      const field = FieldDimension.parseMBQL(dimension.column.field_ref);
+      if (field && field.temporalUnit()) {
+        currentGranularity = formatBucketing(
+          field.temporalUnit(),
+        ).toLowerCase();
+        return true;
+      }
+    }
+  });
+
+  let newGranularity;
+  drilldown.breakouts.some(breakout => {
+    const field = FieldDimension.parseMBQL(breakout);
+    if (field && field.temporalUnit()) {
+      newGranularity = formatBucketing(field.temporalUnit()).toLowerCase();
+      return true;
+    }
+  });
+
+  if (currentGranularity && newGranularity) {
+    return t`See this ${currentGranularity} by ${newGranularity}`;
+  }
+
+  return t`Zoom in`;
 };
 
 export function zoomDrillQuestion(
